refactor(18960): refactor FSM render to isolate internal components
